refactor(bot): add BasePaginationView
- Add BasePaginationView class as a View template for pagination.
