+++ fazbot/bot/view/worldlist_view.py
@@ -1,13 +1,12 @@
 from __future__ import annotations
 
-from datetime import timedelta, timezone
+from datetime import timezone
 from typing import TYPE_CHECKING, Any, Literal, override
 
-from nextcord import ButtonStyle, Color, Embed
-from nextcord.ui import Button, button
+from nextcord import Color, Embed
 from tabulate import tabulate
 
-from fazbot.bot.view._base_view import BaseView
+from fazbot.bot.view._base_pagination_view import BasePaginationView
 from fazbot.bot.view._pagination_embed import PaginationEmbed
 from fazbot.bot.view._view_utils import ViewUtils
 
@@ -17,7 +16,7 @@ if TYPE_CHECKING:
     from fazbot.bot.bot import Bot
 
 
-class WorldlistView(BaseView):
+class WorldlistView(BasePaginationView):
 
     def __init__(
         self,
@@ -40,6 +39,7 @@ class WorldlistView(BaseView):
         )
         await self._interaction.send(embed=self._get_embed_page(1), view=self)
 
+    @override
     def _get_embed_page(self, page: int) -> Embed:
         embed = self._embed.get_base()
         embed.current_page = page
@@ -69,54 +69,3 @@ class WorldlistView(BaseView):
         )
         embed.finalize()
         return embed
-
-    @staticmethod
-    def _timedelta_to_string(td: timedelta) -> str:
-        hours, remainder = divmod(int(td.total_seconds()), 3600)
-        minutes = remainder // 60
-        ret = (str(hours) + "h " if hours > 0 else "") + str(minutes) + "m"
-        return ret
-
-    @button(style=ButtonStyle.blurple, emoji="⏮️")
-    async def first_page_callback(
-        self, button: Button[Any], interaction: Interaction[Any]
-    ) -> None:
-        await interaction.response.defer()
-        self._embed.current_page = 1
-        embed = self._get_embed_page(self._embed.current_page)
-        await interaction.edit_original_message(embed=embed)
-
-    @button(style=ButtonStyle.blurple, emoji="◀️")
-    async def previous_page_callback(
-        self, button: Button[Any], interaction: Interaction[Any]
-    ) -> None:
-        await interaction.response.defer()
-        self._embed.current_page -= 1
-        if self._embed.current_page == 0:
-            self._embed.current_page = self._embed.page_count
-        embed = self._get_embed_page(self._embed.current_page)
-        await interaction.edit_original_message(embed=embed)
-
-    @button(style=ButtonStyle.red, emoji="⏹️")
-    async def stop_(self, button: Button[Any], interaction: Interaction[Any]) -> None:
-        await self.on_timeout()
-
-    @button(style=ButtonStyle.blurple, emoji="▶️")
-    async def next_page(
-        self, button: Button[Any], interaction: Interaction[Any]
-    ) -> None:
-        await interaction.response.defer()
-        self._embed.current_page += 1
-        if self._embed.current_page == (self._embed.page_count + 1):
-            self._embed.current_page = 1
-        embed = self._get_embed_page(self._embed.current_page)
-        await interaction.edit_original_message(embed=embed)
-
-    @button(style=ButtonStyle.blurple, emoji="⏭️")
-    async def last_page(
-        self, button: Button[Any], interaction: Interaction[Any]
-    ) -> None:
-        await interaction.response.defer()
-        self._embed.current_page = self._embed.page_count
-        embed = self._get_embed_page(self._embed.current_page)
-        await interaction.edit_original_message(embed=embed)
